+++ pom.xml
@@ -26,12 +26,4 @@
             </plugin>
         </plugins>
     </build>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.junit.jupiter</groupId>
-            <artifactId>junit-jupiter</artifactId>
-            <version>5.6.0-M1</version>
-        </dependency>
-    </dependencies>
 </project>
