--- src/main/java/org/jalasoft/AccountOrigin.java
@@ -1,9 +1,0 @@
-package org.jalasoft;
-
-/**
- * AccountOrigin
- */
-public enum AccountOrigin {
-    LOCAL,
-    FOREIGN
-}
