+++ src/main/java/org/jalasoft/BankClient.java
@@ -9,20 +9,19 @@ public class BankClient {
     
     private Scanner scanner;
     private boolean done;
-    private BankService bankServices;
+    private Bank bank;
 
     private int currentAccount;
 
     public BankClient() {
         done = false;
-        currentAccount = 0;
     }
 
     /**
-     * Asks the users the command they want to execute and executes that command.
+     * Ask the users the command they want to execute and execute that command 
      */
     public void run() {
-        bankServices = new BankService();
+        bank = new Bank();
         scanner = new Scanner(System.in);
         while (!done) {
             System.out.print("Enter command (0=quit, 1=new, 2=select, 3=deposit, 4=loan, 5=show, 6=interest): ");
@@ -33,9 +32,8 @@ public class BankClient {
     }
 
     /**
-     * Decides what command will be executed based on the command number.
-     * 
-     * @param commandNumber the command that will be executed.
+     * Decides what command will be executed based on the command number
+     * @param commandNumber the command that will be executed
      */
     private void processCommand(int commandNumber) {
         if (commandNumber == 0)
@@ -62,7 +60,7 @@ public class BankClient {
     }
 
     private void newAccount() {
-        int accountNumber = bankServices.newAccount(); //delegates the work to the appropriate class
+        int accountNumber = bank.newAccount(); //delegates the work to the appropriate class
 
         // Own logic
 
@@ -73,14 +71,14 @@ public class BankClient {
     private void select() {
         System.out.print("Enter account#: ");
         currentAccount = scanner.nextInt();
-        int balance = bankServices.getBalance(currentAccount); //delegates the work to the appropriate class
+        int balance = bank.getBalance(currentAccount); //delegates the work to the appropriate class
         System.out.println("The balance of account " + currentAccount + " is " + balance);
     }
 
     private void deposit() {
         System.out.print("Enter deposit amount: ");
         int amount = scanner.nextInt();
-        bankServices.deposit(currentAccount, amount); //delegates the work to the appropriate class
+        bank.deposit(currentAccount, amount); //delegates the work to the appropriate class
     }
 
     private void authorizeLoan() {
@@ -88,17 +86,17 @@ public class BankClient {
 
         int loanAmount = scanner.nextInt();
 
-        if (bankServices.authorizeLoan(currentAccount, loanAmount))
+        if (bank.authorizeLoan(currentAccount, loanAmount))
             System.out.println("Your loan is approved");
         else
             System.out.println("Your loan is denied");
     }
 
     private void showAll() {
-        System.out.println(bankServices.toString());
+        System.out.println(bank.toString());
     }
 
     private void addInterest() {
-        bankServices.payInterest();
+        bank.payInterest();
     }
 }
